refactor: rename platform-bible /internal export to /experimental
Rename the unstable secondary entry point on both platform-bible-react and
platform-bible-utils from `/internal` to `/experimental`. Updates the source
entry files, package.json exports + typesVersions, vite and
dts-bundle-generator build configs, the Storybook source-resolution alias, and
all consumer imports across extensions. Regenerated dist artifacts.
