+++ extensions/src/platform-enhanced-resources/src/components/dictionary-tab/dictionary-tab.component.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { Button } from 'platform-bible-react';
-import { SourceLanguageIndexedList, type IndexedListItem } from 'platform-bible-react/internal';
+import { SourceLanguageIndexedList, type IndexedListItem } from 'platform-bible-react/experimental';
 import type { LocalizedStringValue } from 'platform-bible-utils';
 import { formatReplacementString } from 'platform-bible-utils';
 import {
