+++ lib/platform-bible-react/dts-bundle-generator.config.cts
@@ -20,11 +20,11 @@ const config = {
       },
     },
     {
-      filePath: './src/internal.ts',
-      outFile: './dist/internal.d.ts',
+      filePath: './src/experimental.ts',
+      outFile: './dist/experimental.d.ts',
       noCheck: false,
       libraries: {
-        // `clsx` is inlined here (unlike the index entry) because the internal entry surfaces
+        // `clsx` is inlined here (unlike the index entry) because the experimental entry surfaces
         // cva-typed components (e.g. ProjectSelector/LinkedScrRefButton) without also exporting a
         // symbol that references `clsx`'s `ClassValue` directly. Without inlining clsx, the bundler
         // emits a dangling `CLSX.ClassValue` reference and the generated-bundle check fails.
