+++ lib/platform-bible-react/src/experimental.ts
@@ -1,6 +1,6 @@
 // Unstable API surface for platform-bible-react.
 //
-// Everything exported here is importable from `platform-bible-react/internal`
+// Everything exported here is importable from `platform-bible-react/experimental`
 // but carries NO stability guarantee — symbols may change shape, move, or
 // disappear at any time without a deprecation cycle. Extensions may use
 // it, but should accept the maintenance burden of breaking changes.
